+++ gitflow-completion.sh
@@ -1,18 +1,13 @@
 #!/bin/bash
 
 _completion() {
-	local cur prev opts _ret=1
+	local cur prev opts
 	COMPREPLY=()
 	cur="${COMP_WORDS[COMP_CWORD]}"
 	prev="${COMP_WORDS[COMP_CWORD-1]}"
 	let cword=COMP_CWORD-1
-	let _ret && _ret=0
-	opts="in-progress resolved deployable closed pr"
+	opts="in-progress to-staging resolved deployable closed pr"
 	self_opts="update"
-
-	if [[ $WF_DEBUG -eq 1 ]]; then
-		echo $cur :: $prev :: $cword :: $ret >> aa.log
-	fi
 
 	if [[ ${cword} -eq 0 ]]; then
 		if [[ $cur == *origin* ]]; then
@@ -28,6 +23,8 @@ _completion() {
 			fi
 		elif [[ $prev == "resolved" ]]; then
 			COMPREPLY=( $(compgen -W "sync" -- "${cur}") )
+		elif [[ $prev == "to-staging" ]]; then
+			COMPREPLY=( $(compgen -W "-m" -- "${cur}") )
 		elif [[ $prev == "sync" && $cword == 3 ]]; then
 			COMPREPLY=( $(compgen -W "-m" -- "${cur}") )
 		fi